Updated helper classes to use int rather than uint.
Updated programs to use ints.

+++ HelperClasses.UnitTests/PrimesTests.cs
@@ -10,7 +10,7 @@ namespace HelperClasses.UnitTests
 		[Test]
 		public void GetPrimes_ValidMaxValue_ReturnsCorrectArrayOfValues()
 		{
-			uint[] result = Primes.GetPrimes(5);
+			int[] result = Primes.GetPrimes(5);
 
 			Assert.That(result, Is.EquivalentTo(new int[] { 1, 2, 3, 5 }));
 		}
@@ -18,26 +18,26 @@ namespace HelperClasses.UnitTests
 		[Test]
 		public void GetPrimes_ValidMaxValue_ReturnsCorrectSizedArray()
 		{
-			uint[] result = Primes.GetPrimes(5);
+            int[] result = Primes.GetPrimes(5);
 
 			Assert.That(result.Length, Is.EqualTo(4));
 		}
 
 		[Test]
-		[TestCase((uint)0)]
-		[TestCase((uint)int.MaxValue)]
-		public void GetPrimes_InvalidValues_ThrowsArgumentOutOfRangeException(uint maxValue)
+		[TestCase(0)]
+		[TestCase(int.MaxValue)]
+		public void GetPrimes_InvalidValues_ThrowsArgumentOutOfRangeException(int maxValue)
 		{
 			Assert.Throws<ArgumentOutOfRangeException>(() => Primes.GetPrimes(maxValue));
 		}
 
 		[Test]
-		[TestCase((uint)1, false)]
-		[TestCase((uint)2, true)]
-		[TestCase((uint)3, true)]
-		[TestCase((uint)97, true)]
-		[TestCase((uint)98, false)]
-		public void IsPrime_ValidValues_ReturnsIfValueIsPrime(uint testValue, bool expectedResult)
+		[TestCase(1, false)]
+		[TestCase(2, true)]
+		[TestCase(3, true)]
+		[TestCase(97, true)]
+		[TestCase(98, false)]
+		public void IsPrime_ValidValues_ReturnsIfValueIsPrime(int testValue, bool expectedResult)
 		{
 			bool result = Primes.IsPrime(testValue);
